Move component.js to app path
